feat: Import new <QuestionCaptions/> in <GameHeader/>
- Refactor long template into new reusable <QuestionCaptions/> component
